Time: 30 ms (59.25%), Space: 44.2 MB (38.1%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -8,7 +8,7 @@ typedef unsigned long long ll;
         for (int i = 1; i <= n; i++) {
             for (int j = 1; j <= m; j++) {
                 if (s[i-1] == t[j-1]) {
-                    dp[i][j] += (ll)dp[i-1][j-1]+dp[i-1][j];
+                    dp[i][j] +=dp[i-1][j-1]+dp[i-1][j];
                 }else dp[i][j] = dp[i-1][j];
             }
         }
